MDS-373 | Aligned domain model to new spec
